backend-indexer: refactor follow viem structure
