+++ packages/backend-indexer/package.json
@@ -37,10 +37,10 @@
       "require": "./lib/cjs/actions/index.js",
       "import": "./lib/esm/actions/index.js"
     },
-    "./actions/*": {
-      "types": "./lib/types/actions/*.d.ts",
-      "require": "./lib/cjs/actions/*.js",
-      "import": "./lib/esm/actions/*.js"
+    "./clients": {
+      "types": "./lib/types/clients/index.d.ts",
+      "require": "./lib/cjs/clients/index.js",
+      "import": "./lib/esm/clients/index.js"
     }
   },
   "scripts": {
