+++ packages/backend-indexer/src/actions/index.ts
@@ -1,13 +1,2 @@
-export * from "./erc20.js";
-export * from "./erc721.js";
-export * from "./erc1155.js";
-
-export * from "./getBlock.js";
-export * from "./getBytecode.js";
-export * from "./getLogs.js";
-export * from "./getLogsDecoded.js";
-export * from "./getTransaction.js";
-export * from "./getTransactionReceipt.js";
-export * from "./getTransactionReceiptDecoded.js";
-
-export * from "./updateStateForLog.js";
+export * from "./indexer/index.js";
+export * from "./public/index.js";
