+++ packages/backend-indexer/src/actions/indexer/index.ts
@@ -1,0 +1,4 @@
+export * from "./erc20.js";
+export * from "./erc721.js";
+export * from "./erc1155.js";
+export * from "./updateStateForLog.js";
