+++ packages/backend-indexer/src/actions/public/index.ts
@@ -1,0 +1,7 @@
+export * from "./getBlock.js";
+export * from "./getBytecode.js";
+export * from "./getLogs.js";
+export * from "./getLogsDecoded.js";
+export * from "./getTransaction.js";
+export * from "./getTransactionReceipt.js";
+export * from "./getTransactionReceiptDecoded.js";
