+++ packages/backend-indexer/src/clients/decorators/index.ts
@@ -1,0 +1,1 @@
+export * from "./indexer.js";
