+++ packages/backend-indexer/src/clients/index.ts
@@ -1,0 +1,3 @@
+export * from "./decorators/index.js";
+
+export * from "./createIndexerClient.js";
