+++ packages/backend-indexer/src/index.ts
@@ -1,6 +1,7 @@
+//TODO: Once updateStateForLog removed from controllers/
+// export * from "./actions/index.js";
+export * from "./clients/index.js";
+
 export * from "./createIndexerBackend.js";
 export * from "./isPublicRpcMethod.js";
-
 export * from "./controllers/index.js";
-
-export * from "./createIndexerClient.js";
